backend/requirements.txt
django>=5.0,<5.2
djangorestframework>=3.14
django-cors-headers>=4.3
whitenoise>=6.6
gunicorn>=21.2

# PDF processing
pdfplumber>=0.11

# Embeddings + Vector store
sentence-transformers>=3.0
chromadb>=0.5.3
numpy>=2.0

# LLM
groq>=0.9

# Utils
python-dotenv>=1.0
Pillow>=10.4
